Removing unused classes
