+++ src/test/java/org/fogbow/federatednetwork/ApplicationFacadeTest.java
@@ -1,42 +1,24 @@
 package org.fogbow.federatednetwork;
 
-import org.fogbow.federatednetwork.controllers.FederateComputeUtil;
-import org.fogbow.federatednetwork.controllers.FederatedNetworkController;
-import org.fogbow.federatednetwork.exceptions.FederatedComputeNotFoundException;
-import org.fogbow.federatednetwork.exceptions.NotEmptyFederatedNetworkException;
-import org.fogbow.federatednetwork.exceptions.SubnetAddressesCapacityReachedException;
-import org.fogbow.federatednetwork.model.FederatedComputeInstance;
-import org.fogbow.federatednetwork.model.FederatedComputeOrderOld;
-import org.fogbow.federatednetwork.model.FederatedNetwork;
-import org.fogbowcloud.manager.core.exceptions.InvalidParameterException;
-import org.fogbowcloud.manager.core.exceptions.UnauthenticatedUserException;
-import org.fogbowcloud.manager.core.models.instances.ComputeInstance;
-import org.fogbowcloud.manager.core.models.instances.InstanceState;
 import org.fogbowcloud.manager.core.models.orders.ComputeOrder;
 import org.fogbowcloud.manager.core.models.orders.UserData;
 import org.fogbowcloud.manager.core.models.tokens.FederationUser;
 import org.junit.After;
 import org.junit.Before;
-import org.junit.Test;
 import org.mockito.Mockito;
 
 import java.io.File;
-import java.io.IOException;
 import java.util.*;
 
 import static org.mockito.Mockito.*;
-
-import static org.junit.Assert.*;
 
 public class ApplicationFacadeTest {
 
 	private static final String TEST_DATABASE_FILE_PATH = "federated-networks-test.db";
 
-	private FederatedNetworkController federatedNetworkController;
-
 	private String fakeId = "fakeId";
 
-	@Before
+	/*@Before
 	public void setUp(){
 		deleteTestFiles();
 
@@ -54,7 +36,7 @@ public class ApplicationFacadeTest {
 
 		ApplicationFacade.getInstance().setFederatedNetworkController(federatedNetworkController);
 	}
-
+*/
 	@After
 	public void clean() {
 		deleteTestFiles();
