--- src/test/java/org/fogbow/federatednetwork/ComputeTest.java
@@ -1,18 +1,0 @@
-package org.fogbow.federatednetwork;
-
-import org.junit.Before;
-import org.junit.Test;
-
-public class ComputeTest {
-
-	@Before
-	public void setUp(){
-
-	}
-
-	@Test
-	public void testPostCompute(){
-
-	}
-
-}
